clean stuff up and shift the helpers out to the MMDLib side of things where they rightfully belong
